split token and user into separate services; rename "VerificationToken" classes to just "Token" classes
