+++ backend/src/main/java/com/main/backend/features/token/api/TokenRepository.java
@@ -1,0 +1,7 @@
+package com.main.backend.features.token.api;
+
+import com.main.backend.features.token.entity.TokenEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TokenRepository extends JpaRepository<TokenEntity, Long> {
+}
