+++ backend/src/main/java/com/main/backend/features/token/domain/TokenType.java
@@ -1,0 +1,6 @@
+package com.main.backend.features.token.domain;
+
+public enum TokenType {
+    REGISTRATION,
+    FORGET_PASSWORD
+}
