+++ backend/src/main/java/com/main/backend/features/user/api/UserController.java
@@ -1,6 +1,9 @@
-package com.main.backend.features.user.api.verification;
+package com.main.backend.features.user.api;
 
+import com.main.backend.features.token.api.TokenService;
+import com.main.backend.features.user.domain.User;
 import com.main.backend.features.user.dto.UserDTO;
+import com.main.backend.features.user.entity.UserEntity;
 import com.main.backend.features.user.exception.UserNotFountException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,14 +11,18 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
+
 @Slf4j
 @RestController
-public class VerificationTokenController {
-    private final VerificationTokenService service;
+public class UserController {
+    private final UserService userService;
+    private final TokenService tokenService;
 
     @Autowired
-    public VerificationTokenController(VerificationTokenService service) {
-        this.service = service;
+    public UserController(UserService userService, TokenService tokenService) {
+        this.userService = userService;
+        this.tokenService = tokenService;
     }
 
     @PostMapping("/registration")
@@ -24,12 +31,15 @@ public class VerificationTokenController {
         String response;
 
         try {
+            final UserEntity user = userService.createUser(userDTO.getUsername(), userDTO.getPassword(), userDTO.getEmail());
             status = HttpStatus.OK;
-            response = service.registration(userDTO.getUsername(), userDTO.getPassword(), userDTO.getEmail());
+            response = tokenService.registration(user);
+
             log.info("Registration token created successfully");
         } catch (Exception ex) {
             status = HttpStatus.BAD_REQUEST;
             response = "Error during token creation!";
+
             log.error(response, ex);
         }
 
@@ -38,7 +48,7 @@ public class VerificationTokenController {
 
     @GetMapping("/registration/{token}")
     public ResponseEntity<String> confirmRegistration(@PathVariable String token) {
-        return ResponseEntity.ok(service.confirmRegistration(token));
+        return ResponseEntity.ok(tokenService.confirmRegistration(token));
     }
 
     @PostMapping("/forgot-password")
@@ -47,15 +57,28 @@ public class VerificationTokenController {
         String response;
 
         try {
+            final UserEntity user = userService.getUserEntity(userId);
+            if (user == null) {
+                log.error("User: {} not found", userId);
+                throw new UserNotFountException("User not found");
+            }
+
             status = HttpStatus.OK;
-            response = service.forgotPassword(userId);
+            response = tokenService.forgotPassword(user);
+
             log.info("Forget-password token created successfully");
         } catch (Exception ex) {
             status = HttpStatus.BAD_REQUEST;
             response = "Error during token creation!";
+
             log.error(response, ex);
         }
 
         return ResponseEntity.status(status).body(response);
+    }
+
+    @GetMapping("/users")
+    public ResponseEntity<List<User>> getUsers() {
+        return ResponseEntity.ok(userService.getUsers());
     }
 }
