+++ backend/src/main/java/com/main/backend/features/user/api/UserRepository.java
@@ -1,4 +1,4 @@
-package com.main.backend.features.user.api.userdata;
+package com.main.backend.features.user.api;
 
 import com.main.backend.features.user.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
