--- backend/src/main/java/com/main/backend/features/user/api/userdata/UserController.java
@@ -1,24 +1,0 @@
-package com.main.backend.features.user.api.userdata;
-
-import com.main.backend.features.user.domain.User;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-public class UserController {
-    private final UserService service;
-
-    @Autowired
-    public UserController(UserService service) {
-        this.service = service;
-    }
-
-    @GetMapping("/users")
-    public ResponseEntity<List<User>> getUsers() {
-        return ResponseEntity.ok(service.getUsers());
-    }
-}
