--- backend/src/main/java/com/main/backend/features/user/api/verification/VerificationTokenRepository.java
@@ -1,7 +1,0 @@
-package com.main.backend.features.user.api.verification;
-
-import com.main.backend.features.user.entity.VerificationTokenEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface VerificationTokenRepository extends JpaRepository<VerificationTokenEntity, Long> {
-}
