--- backend/src/main/java/com/main/backend/features/user/domain/VerificationTokenType.java
@@ -1,6 +1,0 @@
-package com.main.backend.features.user.domain;
-
-public enum VerificationTokenType {
-    REGISTRATION,
-    FORGET_PASSWORD
-}
